requirements.txt
streamlit
tensorflow
numpy
matplotlib
pylsl
scipy
mne
muselsl
openai==0.28
python-dotenv